feat: add query subcommand
The `query` subcommand can be used to print information about all
packages in a monorepo to avoid needing to invoke `lerna list`
repeatedly.

For example, instead of invoking

```
lerna list --all --json --scope=library-a --include-dependencies
lerna list --all --json --scope=library-b --include-dependencies
lerna list --all --json --scope=library-c --include-dependencies
...
```

we can use `query` to return all of this information at once

```
monorepo query internal-dependencies --format path
```

where `format` is either `path` or `name` (the default, mirroring
lerna's default behavior).

This sub-command is mostly experimental and a work in progress; it is
not a part of the official spec nor the TypeScript implementation of the
typescript-tools. It is a perfect addition to the Rust implementation
beacuse the requirement is gathering this information as quickly as
possible.